- Switched application from yml to properties type. - Moved the Application into a new package.
Kristóf Juhász <killerammer13@gmail.com>
Ádám Békési <bekesiadam@gmail.com>
