+++ booksite/src/main/java/hu/nye/webapp/books/Application.java
@@ -1,3 +1,5 @@
+package hu.nye.webapp.books;
+
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
