+++ booksite/src/main/java/hu/nye/webapp/books/entity/Book.java
@@ -1,25 +1,27 @@
-package dto;
+package hu.nye.webapp.books.entity;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 
-import javax.validation.constraints.NotNull;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 import java.util.Date;
 
-@NoArgsConstructor
 @Getter
 @Setter
-@ToString
-public class BookDTO {
-
+@NoArgsConstructor
+@Entity
+public class Book {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    @NotNull
     private String title;
     private String author;
     private Date releaseDate;
     private int price;
-    @NotNull
     private int pageNumber;
 }
