+++ booksite/src/main/java/hu/nye/webapp/books/exception/BookNotFoundException.java
@@ -1,4 +1,4 @@
-package exception;
+package hu.nye.webapp.books.exception;
 
 public class BookNotFoundException extends  RuntimeException{
