+++ booksite/src/main/java/hu/nye/webapp/books/exception/InvalidBookException.java
@@ -1,4 +1,4 @@
-package exception;
+package hu.nye.webapp.books.exception;
 
 import java.util.List;
