+++ booksite/src/main/java/hu/nye/webapp/books/repository/BookRepository.java
@@ -1,6 +1,6 @@
-package repository;
+package hu.nye.webapp.books.repository;
 
-import entity.Book;
+import hu.nye.webapp.books.entity.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BookRepository extends JpaRepository <Book, Long>{
