--- booksite/src/main/resources/application.yml
@@ -1,9 +1,0 @@
-spring:
-  datasource:
-    url: jdbc:mysql://localhost:3306/booksite
-    username: root
-    password:
-
-springdoc:
-  swagger-ui:
-    path: /swagger-ui.html
